+++ Example/Core/Sources/Core/L10n.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum L10n {
+    static let condensedText = NSLocalizedString("condensed.text", bundle: Bundle.module, comment: "")
+    static let dummyTextGerman = NSLocalizedString("dummy.text.german", bundle: Bundle.module, comment: "")
+    static let monospacedText = NSLocalizedString("monospaced.text", bundle: Bundle.module, comment: "")
+    static let pangramArabic = NSLocalizedString("pangram.arabic", bundle: Bundle.module, comment: "")
+    static let pangramEnglish = NSLocalizedString("pangram.english", bundle: Bundle.module, comment: "")
+    static let pangramGreek = NSLocalizedString("pangram.greek", bundle: Bundle.module, comment: "")
+    static let pangramHebrew = NSLocalizedString("pangram.hebrew", bundle: Bundle.module, comment: "")
+}