+++ Sources/NSAttributedStringBuilder/Extensions/AttributedStringBuilding+macOS.swift
@@ -15,6 +15,18 @@ public extension AttributedStringBuilding {
         addingAttribute(.font, value: fontWithTrait(.italic))
     }
 
+    /// Modifies the font of the text to use a fixed-width variant of the current font, if possible.
+    /// - Returns: A copy of the modified attributed string.
+    func monospaced() -> NSAttributedString {
+        addingAttribute(.font, value: fontWithTrait(.monoSpace))
+    }
+
+    /// Adds the condensed trait to the current font.
+    /// - Returns: A copy of the modified attributed string.
+    func condensed() -> NSAttributedString {
+        addingAttribute(.font, value: fontWithTrait(.condensed))
+    }
+
     /// Adds an image at the end of the attributed string.
     /// - Parameters:
     ///   - image: The image that will be added to the string.